DON-713: Refactor: extract Highlight cards component from home component
Also removed the appear-at and disappear-at properties of the highlight
card that as they are now always set to "asap" and "never" respectively
and therefore useless
