+++ src/app/highlight-cards/HighlightCard.ts
@@ -1,17 +1,6 @@
 import { brandColour } from "@biggive/components/dist/types/globals/brand-colour"
 
 export type HighlightCard = {
-  /**
-   * If not 'asap', the card should not be displayed until the date given.
-   * Remember to set the timezone appropriately when creating these date objects.
-   */
-  appearAt: Date | 'asap',
-
-  /**
-   * If not 'never', the card should disappear at this date.
-   */
-  disappearAt: Date | 'never',
-
   backgroundImageUrl: URL,
   // There is ambiguity in the components library about whether brand-6 is grey or turquoise. Best to avoid using brand-6 and
   // use brand-mhf-turquoise instead.
@@ -77,8 +66,6 @@ export const SFAPIHighlightCardToHighlightCard = (experienceUriPrefix: string, b
   };
 
   return {
-    appearAt: sfApiHighlightCard.appearAt,
-    disappearAt: sfApiHighlightCard.disappearAt,
     headerText: sfApiHighlightCard.headerText,
     bodyText: sfApiHighlightCard.bodyText,
     iconColor: campaignFamilyColours[sfApiHighlightCard.campaignFamily] || "primary",
