+++ src/app/highlight-cards/highlight-cards.component.cy.ts
@@ -1,0 +1,7 @@
+import { HighlightCardsComponent } from './highlight-cards.component'
+
+describe('HighlightCardsComponent', () => {
+  it('should mount', () => {
+    cy.mount(HighlightCardsComponent)
+  })
+})
