+++ src/app/highlight-cards/highlight-cards.component.html
@@ -1,0 +1,35 @@
+@if (highlightCards.length === 1) {
+  <div class="single-highlight-card">
+    <biggive-basic-card
+      [spaceBelow]='4'
+      [headingLevel]="2"
+      [buttonColourScheme]="'clear-primary'"
+      [class]="'basic-card-padding'"
+      [iconColour]="highlightCards[0]!.iconColor"
+      [backgroundColour]="highlightCards[0]!.iconColor"
+      [backgroundImageUrl]="highlightCards[0]!.backgroundImageUrl.toString()"
+      [mainTitle]="highlightCards[0]!.headerText"
+      [subtitle]="highlightCards[0]!.bodyText"
+      [buttonLabel]="highlightCards[0]!.button.text"
+      [buttonUrl]="highlightCards[0]!.button.href.toString()"
+    />
+  </div>
+} @else {
+  <biggive-grid [columnCount]="highlightCards.length" column-gap="3">
+    @for (card of highlightCards; track $index) {
+      <biggive-basic-card
+        [spaceBelow]='4'
+        [headingLevel]="2"
+        [buttonColourScheme]="'clear-primary'"
+        [class]="'basic-card-padding'"
+        [iconColour]="card.iconColor"
+        [backgroundColour]="card.iconColor"
+        [backgroundImageUrl]="card.backgroundImageUrl.toString()"
+        [mainTitle]="card.headerText"
+        [subtitle]="card.bodyText"
+        [buttonLabel]="card.button.text"
+        [buttonUrl]="card.button.href.toString()"
+      />
+    }
+  </biggive-grid>
+}
