+++ src/app/home/home.component.html
@@ -12,41 +12,9 @@
       <!-- The grid layout works fine for 2 or three cards but not quite for a single card. Easier to treat single
       card as a special case rather than changing the grid implemnation or replacing it entirely.
       -->
-      @if (highlightCardsToShow.length === 1) {
-        <div class="single-highlight-card">
-          <biggive-basic-card
-            [spaceBelow]='4'
-            [headingLevel]="2"
-            [buttonColourScheme]="'clear-primary'"
-            [class]="'basic-card-padding'"
-            [iconColour]="highlightCardsToShow[0]!.iconColor"
-            [backgroundColour]="highlightCardsToShow[0]!.iconColor"
-            [backgroundImageUrl]="highlightCardsToShow[0]!.backgroundImageUrl.toString()"
-            [mainTitle]="highlightCardsToShow[0]!.headerText"
-            [subtitle]="highlightCardsToShow[0]!.bodyText"
-            [buttonLabel]="highlightCardsToShow[0]!.button.text"
-            [buttonUrl]="highlightCardsToShow[0]!.button.href.toString()"
-            />
-        </div>
-      } @else {
-        <biggive-grid [columnCount]="highlightCardsToShow.length" column-gap="3">
-          @for (card of highlightCardsToShow; track $index) {
-            <biggive-basic-card
-              [spaceBelow]='4'
-              [headingLevel]="2"
-              [buttonColourScheme]="'clear-primary'"
-              [class]="'basic-card-padding'"
-              [iconColour]="card.iconColor"
-              [backgroundColour]="card.iconColor"
-              [backgroundImageUrl]="card.backgroundImageUrl.toString()"
-              [mainTitle]="card.headerText"
-              [subtitle]="card.bodyText"
-              [buttonLabel]="card.button.text"
-              [buttonUrl]="card.button.href.toString()"
-              />
-          }
-        </biggive-grid>
-      }
+      <app-highlight-cards
+        [highlightCards]="highlightCards">
+      </app-highlight-cards>
     </div>
     <biggive-page-section
       colour-scheme="primary"
